requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.2.2
django-cors-headers==3.7.0
idna==2.10
mysql==0.0.2
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
